Commands, Events, Handler structures

+++ events/message.js
@@ -1,6 +1,8 @@
 class Event {
   constructor (client) {
     this.client = client
+    this.name = 'message'
+    this.listener = (...args) => this.run(...args)
     this.classPrefix = '[Events:onMessage'
     this.defaultPrefix = {
       handleCommand: `${this.classPrefix}:handleCommand]`
@@ -20,19 +22,16 @@ class Event {
     if (message.channel.type === 'dm') return message.channel.send(`${this.client._options.constructors.EMOJI_NO}  DM 에서는 명령어를 사용하실수 없어요..\n${this.client._options.constructors.EMOJI_NO}  You can\'t use commands on the DM.`)
     if (message.guild && !message.member) await message.guild.fetchMember(message.author)
     if (!this.client.utils.permissionChecker.checkChannelPermission(message.guild.me, message.channel, ['SEND_MESSAGES'])) return
-    // Test
-    this.client.redisClient.publish('asdf', JSON.stringify({ message: message.content }))
-    // End of Test
     await this.client.database.checkGuild(message.guild.id)
     await this.client.database.checkMember(message.member.id, message.guild.id)
     await this.client.database.checkUser(message.author.id)
     const prefix = this.client._options.bot.prefix
     if (message.content.startsWith(prefix)) {
-      const guildData = await this.client.database.getGuild(message.guild.id)
-      const { locale } = guildData
-      const picker = this.client.utils.localePicker
       if (message.author.awaitQuestion) return
+      const picker = this.client.utils.localePicker
+      const guildData = await this.client.database.getGuild(message.guild.id)
       const userData = await this.client.database.getUser(message.author.id)
+      const { locale } = guildData
       const args = message.content.slice(prefix.length).trim().split(/ +/g)
       const command = args.shift().toLowerCase()
       if (userData.blacklisted && !this.client._options.bot.owners.includes(message.author.id)) return this.client.logger.warn(`${this.defaultPrefix.handleCommand} Blacklisted User Issued Command ${command}, [${args.join(', ')}]`)
@@ -52,48 +51,43 @@ class Event {
         command: command,
         userPermissions: userPermissions
       })
-      const Command = this.client.commands.get(command) || this.client.commands.get(this.client.aliases.get(command))
-      if (Command) {
+      const commandClass = this.client.commands.get(command) || this.client.commands.get(this.client.aliases.get(command))
+      if (commandClass) {
         if (this.client.shuttingDown) return message.channel.send(picker.get(locale, 'UNABLE_USE_COMMAND_SHUTDOWN'))
-        let ablePermissions = 0
         if (this.client.chkRightChannel(message.channel, guildData.tch)) {
-          if (Command.command.require_nodes) {
-            if (guildData.tch === '0') this.client.audio.textChannels.set(message.guild.id, message.channel.id)
-            if (!this.client.audio.getNode()) return message.channel.send(picker.get(locale, 'AUDIO_NO_NODES'))
-          }
-          if (Command.command.require_playing) {
-            if (!this.client.audio.players.get(message.guild.id)) return message.channel.send(picker.get(locale, 'AUDIO_REQUIRE_PLAYING'))
-          }
-          if (Command.command.require_voice) {
-            const vch = guildData.vch
-            if (!message.member.voice.channel) return message.channel.send(picker.get(locale, 'AUDIO_JOIN_VOICE_FIRST'))
-            if (this.client.audio.utils.getVoiceStatus(message.member).listen === false) return message.channel.send(picker.get(locale, 'AUDIO_LISTEN_PLEASE'))
-            if ((message.guild.me.voice.channelID && message.guild.me.voice.channelID !== message.member.voice.channelID)) return message.channel.send(picker.get(locale, 'AUDIO_SAME_VOICE', { VOICECHANNEL: message.guild.me.voice.channelID }))
+          /**
+           * Requirements
+           */
+          const { requirements } = commandClass
+          const { audioNodes, playingStatus, voiceStatus } = requirements
+          // Requirements - audioNodes, playingStatus
+          if (audioNodes && !this.client.audio.getNode()) return message.channel.send(picker.get(locale, 'AUDIO_NO_NODES'))
+          if (audioNodes && guildData.tch === '0') this.client.audio.textChannels.set(message.guild.id, message.channel.id)
+          if (playingStatus && !this.client.audio.players.get(message.guild.id)) return message.channel.send(picker.get(locale, 'AUDIO_REQUIRE_PLAYING'))
+          // Requirements.voiceStatus
+          if (voiceStatus) {
+            const { listenStatus, sameChannel, voiceIn } = voiceStatus
+            const { vch } = guildData
+            if (voiceIn && !message.member.voice.channel) return message.channel.send(picker.get(locale, 'AUDIO_JOIN_VOICE_FIRST'))
             if ((message.guild.me.voice.channelID && message.guild.me.voice.channelID !== message.member.voice.channelID) && !this.client.chkRightChannel(message.member.voice.channel, vch)) return message.channel.send(picker.get(locale, 'AUDIO_NOT_DEFAULT_CH', { VOICECHANNEL: vch }))
+            // Requirements.voiceStatus - listenStatus
+            if (listenStatus && this.client.audio.utils.getVoiceStatus(message.member).listen === false) return message.channel.send(picker.get(locale, 'AUDIO_LISTEN_PLEASE'))
+            // Requirements.voiceStatus - sameChannel
+            if (sameChannel && (message.guild.me.voice.channelID && message.guild.me.voice.channelID !== message.member.voice.channelID)) return message.channel.send(picker.get(locale, 'AUDIO_SAME_VOICE', { VOICECHANNEL: message.guild.me.voice.channelID }))
           }
-          for (const userPerm of userPermissions) {
-            if (Command.command.permissions.includes(userPerm)) {
-              ablePermissions++
-              this.client.logger.debug(`${this.defaultPrefix.handleCommand} (${message.channel.id}, ${message.id}, ${message.author.id}) Treating command ${Command.command.name} at ${new Date().getTime()}`)
-              try {
-                await Command.run(compressed)
-                message.author.awaitQuestion = false
-                break
-              } catch (e) {
-                message.author.awaitQuestion = false
-                if (e instanceof this.client.utils.errors.PermError) return message.channel.send(picker.get(locale, 'ERROR_PERMISSION', { PERMS: e.perms.join(', ') }))
-                this.client.logger.error(`${this.defaultPrefix.handleCommand} (${message.channel.id}, ${message.id}, ${message.author.id}) Unexpected Error: ${e.name}: ${e.stack}`)
-                await message.channel.send(picker.get(locale, 'HANDLE_COMMANDS_ERROR', { UUID: this.client.database.addErrorInfo('commandError', e.name, e.stack, message.author.id, message.guild.id, Command.command.name, args) }))
-              }
-            }
+          const havePermissions = commandClass.permissions.filter(el => userPermissions.includes(el))
+          this.client.logger.debug(`${this.defaultPrefix.handleCommand} (${message.channel.id}, ${message.id}, ${message.author.id}) Treating command ${commandClass.name} at ${new Date().getTime()}`)
+          if (havePermissions.length === 0) return message.channel.send(picker.get(locale, 'HANDLE_COMMANDS_NO_PERMISSIONS', { REQUIRED: commandClass.permissions.join(', ') }))
+          try {
+            await commandClass.run(compressed)
+          } catch (e) {
+            if (e instanceof this.client.utils.errors.PermError) return message.channel.send(picker.get(locale, 'ERROR_PERMISSION', { PERMS: e.perms.join(', ') }))
+            this.client.logger.error(`${this.defaultPrefix.handleCommand} (${message.channel.id}, ${message.id}, ${message.author.id}) Unexpected Error: ${e.name}: ${e.stack}`)
+            await message.channel.send(picker.get(locale, 'HANDLE_COMMANDS_ERROR', { UUID: this.client.database.addErrorInfo('commandError', e.name, e.stack, message.author.id, message.guild.id, commandClass.name, args) }))
           }
-          if (ablePermissions === 0) {
-            await message.channel.send(picker.get(locale, 'HANDLE_COMMANDS_NO_PERMISSIONS', { REQUIRED: Command.command.permissions.join(', ') }))
-          }
+          message.author.awaitQuestion = false
         } else {
-          if (this.client.utils.permissionChecker.checkChannelPermission(message.guild.me, message.channel, ['MANAGE_MESSAGES'])) {
-            await message.delete()
-          }
+          if (this.client.utils.permissionChecker.checkChannelPermission(message.guild.me, message.channel, ['MANAGE_MESSAGES'])) await message.delete()
           try {
             await message.author.send(picker.get(locale, 'HANDLE_COMMANDS_DEFAULT_TEXT', {
               SERVER: message.guild.name,
@@ -111,11 +105,4 @@ class Event {
   }
 }
 
-module
-  .exports = Event
-
-module
-  .exports
-  .info = {
-    event: 'message'
-  }
+module.exports = Event
